shortid better sitribution, file hash fix
